+++ src/components/AssetValue.tsx
